voice-assistant-platform/backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
structlog==24.4.0
numpy==1.26.4
scikit-learn==1.5.2
transformers==4.46.0
torch==2.5.0
openai-whisper==20240930
TTS==0.22.0
aiofiles==24.1.0
